Rename DocType to Group
We now call group what we called docType before, mostly because it is
not actually a clean separation by js obj type but a top-level grouping
of functionality. The COMPONENT group also contains non-component types.

So we will have doc.group as the top-level group and then have a
separate type enum to reflect that reality.