Add reading position to TOC sidebar
Add a reading poosition context to track the position on the document.
Registering sections to the context allows for the automatic
determination of the most visible (in %), most top in viewport
positioned section to be choosen for marking in the toc sidebar.

Included is a parser to apply the same logic to markdown static pages.