Implement simplified DropdownMenu component
We had strange concurrent update issues with dropdown buttons not
working sometimes because a react update interferes with it and makes
the dropdown code in some cases but consistently think the click on the
button was an "outside" click because the stale element attached as
event target is indeed not a child of the post-update container :\

So war we have been using one dropdownState on the view level to drive
multiple drop downs within the view. This is not optimal from the
updates view point as every opening and closing of the menu triggers a
complete view rerender.

With this we move the dropdown state to the local state of that specific
dropdown. This also seems to fix our update issues.

Just wrap your buttons or dropdown items in this component.