Refactor ViewState
ViewState was a mix of static state artifacts and some dynamic
aspects/the transition map which did not actually work at all because we
would reuse the same transition map over and over and that was
implicitly bound to the process and scope it was initialized with, which
is of course plain wrong.

So we clean this up and make ViewState only contain the static
declarative aspects of the view and keep all transitionMaps in a Map in
process where they're keyed to the state.

We could key by view name, but this way we enforce the view state identity a
bit.