Transition: Fix target setter
Unaware of the existing of the setter, I just assumed t.target = XXX; to
work which it doesn't because the old state name validation cannot
really work anymore.

Theoretically, we could spider along the transition maps to find
possible view states starting from our known start state, but with the
possibility of dynamically setting transition.target to any other view
state, we can't ever be sure to know all states.

As it is now, we just make sure that we initialize the bound transition map
for the local process the first time it enters any view-state.

Simplifies the setter and removes the validation. We could actually
remove the setter, but we keep it for now.