DataGrid: Check computed columns
Checks if a default filtered column is a computed field and throws an
error if that is the case.

To filter or sort computed fields in a datagrid, you have to construct
the same computation as FilterDSL expression for the filter and/or
computed column.

For example if we have a computed column that takes the "name" and the
"num" property of an object and concatenates them with a colon in
between you could do:

    <DataGrid.Column
        name="extra"
        filter={
            val => field("name")
                .concat(value(':'))
                .concat(field('num'))
                .containsIgnoreCase(value(val, 'String'))
        }
        sort={ field("name")
            .concat(value(':'))
            .concat(field('num'))  }
    />