Type Utils: Introduce lookupContext
So far we've got along with a lookupType(typeName, path) function that
resolves the field definition of a given (GraphQLish) path starting at
the a given type.

For some purposes, however, we need both the type name of the parent
object and the field definition.

This changes the lookupType method into a lookupTypeContext method that
returns a typeContext {domainType, field} and changes the original
lookupType to use that and only return the field def part of it.

It introduces an additional object creation which should not matter
because for now, the type is either part of a memoized structure or it
is the result of runtime inspection of a iQuery document, the querying
of which absolutely dwarves the new object creation.

Should this become a measurable problem at some point, we might want to
consider splitting both methods again to trade redundancy for
performance.