GridStateForm: Refactor value initialization.
This changes the algorithm to resolve initial Datagrid filter input
values from the pre-existing iQuery condition filter.

The old algorithm seems to have been broken by the component() changes
following the FilterDSL builder optimization changes, which went
unnoticed because there was no test at all.

I added a test for this behavior to automaton-test at least and will add
an automated test here later.

The new algorithm can handle

 * simple filters
 * simple filteres with multiple values ( between() )
 * filter on computed columns with constant value() nodes that don't
   originate from an user input

Most of the changes deal with the more diverse component() structure
now, the other change is that we don't use the first two values for
between() but the last two.