URLField: Enable clicking link while editing it
So far we have displayed a (disabled) text input for the NORMAL and
READ_ONLY modes and just the URL link for PLAIN_TEXT.

This adds an extra link-addon on the left side of the user input that
can be clicked if the current value matches a minimal URL validation.

It just checks for // to prevent relative links which might lead to bad
behavior beyond just being a wrong link while allowing protocol relative
links, too.

Again we make sure to open the link while protecting ourselves from
anyone trying to exploit the accidental parenting by adding rel="noopener
noreferer".