Process: Fix mobx action storage
There was a pretty serious bug left in the caching of the mobx actions
we create to wrap our raw transition functions.

The key to store them inside the process storage was built from the
source state instance and not its name leading to names like
"[object].transition" which then would be the same for every transition
in different states but with the same name.

This fixes that behavior.

Cleans up the code a bit.