Add form shortcut system
Add a system for a shortcut sidebar to be added to a process for
enhanced interaction with forms.
The sections register inside the context for a sidebar button, which
will then be rendered and, depending on the validation of contained
input elements with fieldcontexts, can show, if errors are present
inside the specific section.
The sidebar will only be rendered, if a section is registered, else it
will be removed from the content flow.

To use this, form sections will need to be inside a <Section> element.
The sidebar needs to be added as <ShortcutSidebar>.