Replace url module
Node's "url" module is deprecated and is also no longer shimmed by
webpack.

This replaces the few locations where we used it with a new util
function using the URLSearchParams API. Regretfully, we really can't use
the Map-ish params object in the code because we need an actual object.

Quick solution is to implement a new searchParams utility function that
returns an object with parameters for the search part of given URL.