Switch to MobX 5
We have some breaking changes for mobx objects. We need to call
makeObservable(this) in their constructor

useLocalObservable replaced useLocalStore

In ChangeDialog there was a problem with some reads happening outside
the <Observer> which did not work. Not sure how it ever worked, but it
did apparently.

Moving the reads into the observer fixes the tests.

Upgrades the tests to new versions of mobx / @testing-library. Adds some
formal initializations to test.