Experimental: Add config.skipIndexTreeCloning
We have a very smelly cloning operations in <Tree.IndexedObjects/> which
once solved very odd update problems. Now I had some problems with it in
the tests and overall would like to get rid of it.

So the current code keeps cloning in <Tree.IndexedObjects/> unless
config.skipIndexTreeCloning is truthy.

The goal would be to be able to test both variants more quickly and then
decide on removing it or not and to remove config.skipIndexTreeCloning
afterwards.