Refactor postprocessors
So far we've interpreted the GraphQL post-processors as type-based
synchronous data-fiddling, but it can be so much more with just a few
changes.

Changes the internal postProcess function so that it handles promises
correctly. A postprocessor can just return a non-promise value as before
but it can also return a promise whose return value is used to update
the result object when it resolves.

postProcess waits for all promises to conclude before resolving to the
modified result object.