Implement promiseUI helper
So far we've been lacking "toast" message like functionality, i.e.
showing pretty little info-popups that inform the user of something
happening and mostly go away automatically.

We wanted to support generic response wrappers for actions that are
automatically translated and that should be project-specific so that the
application author can define their own responses and toasts and extra
fields within the wrapper etc pp.

With the refactored GraphQL postprocessors that most simple
implementation as such an postprocessor would work, but lacks one
crucial feature, which is to be able to show a little spinner toast
first which is then replaced with the result as soon as the fetch ends.

react-toastify's built-in promise helper however does not allow for the
generic wrappers easily.

So the compromise between usability and usefulness is to introduce our
own promise wrapper which we call promiseUI

promiseUI allows to wrap a GraphQL query promise. It displays an
localized loading text and by default ends the request with a
simple localized "Ok" success.

The application author can use a second function `configurePromiseUI`
to define a generic type and a conversion method to produce
react-toastify option objects from the generic data.

Usage
=====

Let's look at the test implementation in automaton-test:

--- generic type -------------------------------------------------------
package de.quinscape.automatontest.model.result;

/**
 * Generic result wrapper. Allows client-side code to register
 * post-processors based on it.
 *
 * @param <P>   payload type
 */
public class Result<P>
{
    private final ResultType type;
    private final P payload;

    public Result(ResultType type, P payload)
    {
        this.type = type;
        this.payload = payload;
    }

    /**
     * Returns the result type.
     *
     * @return result type
     */
    public ResultType getType()
    {
        return type;
    }

    /**
     * Returns the payload of type [P]
     *
     * @return payload
     */
    public P getPayload()
    {
        return payload;
    }
}

--- enum type ----------------------------------------------------------

package de.quinscape.automatontest.model.result;

public enum ResultType
{
    SUCCESS,
    INFO,
    WARNING,
    DEFAULT,
    ERROR,
    SILENT;
}

--- JavaScript config --------------------------------------------------

    configurePromiseUI(
        "de.quinscape.automatontest.model.result.Result",
        result => {

            const { type } = result;

            if (type === "SILENT")
            {
                return null;
            }

            return {
                type,
                render: type,
                autoClose: type === "ERROR" ? false : null
            }
        }
    )

success,info,warning,error and default are the react-toastify styles
which can also be given as uppercase in our converter function.

SILENT is special-cased to return no option object which then leads to
the loading indicator to be just removed.

We configure "autoClose" to be false for ERROR so that the user has to
close these messages and they don't just disappear.