Add controls section to shortcut sidebar
The constrols section for the shortcut sidebar will always be visible,
while all the shortcuts may become scrollable. That way actions like
save, reset, etc will always be accessible no matter where inside the
form you are.
Controls are added simply as children of the sidebar, and the section is
only rendered, if there is at least one child present.