Add inline mode for dropdown menu
The past behaviour of the dropdown was to be always floated to the
right side of the container. The new inline mode alows for it to stay at
the defined position in the layout to be inserted e.g. in a button row.

This change was done for features like an add button with multiple
choices, where you will see a dropdown menu on a seemingly normal
button element containing mentioned choices.