Refactor GraphQL Method aliasing adjacent code
Finishes the GraphQL method aliasing to produce an inverted alias map
like domainql-form requires them now and it changes the former "methods"
array in the query declaration to a methodCalls now containing the
aliased name.

This fixes the big use case of invoking one GraphQL method multiple
times with different parameters and aliasing the results.

The inverted alias map and the methodCalls change also echo through the
code base for every code that touches the query definition to do clever
things with them.

Changes the tests to the new standards.