Implement ConditionEditor
Second attempt at creating an interactive editor for our FilterDSL
expressions. The first sketch in iqed once worked better, but it became
clear that the approach wouldn't lead to an acceptable UI and also it
became apparent that we should not tie the condition error too closely
into the InteractiveQuery mechanism as we are already using the
conditions in other contexts.

Having filter implementations across the full-stack of JavaScript,
server-side Java and even SQL execution, it has become somewhat of a
default go-to for all our data-driven condition needs.

This condition editor provides a general purpose condition editor that
can be integrated into all kinds of contexts, be it interactive query
editing or user-editing of validation rules.

Uses the d3-flextree package to layout two different trees. There is a
simplified general tree that shows all operations inline and the
operation tree we shows them as individual leaves to facilitate
operation editing.

We erroneously added d3-flextree as dependency in 2fe799e6e00317735.
This should have happened here.

Display rotation
================

d3-flextree trees are always oriented so that they grow top to bottom,
but we need them to grow left to right. This means we have to rotate all
tree coordinates coming from d3-flextree 90 degrees counter-clockwise.
But just doing that messes with the expected tab order, so we also need
to mirror the graph vertically.

This complicates the code a bit at some points.

Note that the normal 2D ccw rotation

    x' = y
    y' = -x

then turns into

    x' = y
    y' = x - height

if we apply the y-mirroring which also changes the reference point for
the enclosing rectangle