Add StickyNav for sticky navigation fixes
The new StickyNav compontent will report its height into a context, to
allow the StickySidebars to update their height and top based on the
remaining space. This allows for a much better user experience, since
the height of the navigation can change and its height is no longer
predicted, but instead calculated.
Also the Section Shorttcuts will take advantage of this, too. They do
now properly scroll to the correct height instead of "burrying" the
content under the navigation container.