Fix error handling in registerBigDecimalConverter
The error was only thrown, if the resulting precision resulted to a null
value, but the value would rather be undefined. Strict equality check is
not going to find undefined values if checked for null. Either a loose
equality operator is required (nullsy check) or a cast to boolean for a
falsy check.