Process: Change initialization order
We change the initialization order so that we create the process first
and register it as current process. This way, the deep scope initialization
can access the process already which we need for the working set.

Since the reordering is largely and internal detail with no
collaborators in between it shouldn't be a problem.