Implement incremental WorkingSet
So far, figuring out the actual changes for a working set was done when
the final .merge() call happened and it was a pretty elaborate and slow
process, surely not something we want to do all the time.

But we need to know when changes happened, both to visualize it in the
UI as well as to implement things like browser warnings on dirty pages.

This commit refactors the working set so that it can track the changes
incrementally. A now reduced number of registered objects is now tracked
with dedicated reactions that keep an observable changes map per entry
up to date

To reduce the number of registrations/reactions, the tracking algorithm
was cut short by default. Only the registered objects and it's relations
are tracked. If there are are dependend or embedded objects that are
edited, they need to be registered manually at this point.

The old code used to follow embedded objects and only required the
registering of dependent objects.

In future versions, we might use the new relation meta tags to reenable
the old auto-tracking selectively.

workingSet.isModified and workingSet.hasChanges now fully reflect the
state that will be submitted to the merge process. Both are computed
observables.