Implement dirty tracking
Now that we have made the workingSet reflect its true status, the actual
implementation of the browser protecting our changes is rather trivial.

Adds an onbeforeunload handler based on a isDirty(process) function.

This function uses scope.workingSet.hasChanges if the current scope has
a "workingSet" property.

Otherwise it checks for a isDirty() method on the scope which can be
used to implement other ways that do not use the workingSet
functionality.