Dirty-Checking v2
Refactors the isDirty method to be just "process.isDirty". The method
now not only checks the current process, but all parent processes as
well and returns "is dirty" if any of them is dirty.

Apart from the browser protecting us from unwanted page navigation and
page closing, we still need to protect our internal transitions that are
destructive.

Most of all that is runProcess (and runProcessURI).

Adds a new library method `confirmDestructiveTransition()` which checks
the current process to be dirty and has the user confirm the transition
via a browser confirm dialog "[Delete Dirty Changes?]"

If the user says ok, the method returns true, otherwise false.

Rejects runProcess promise if the user cancels such a destructive
transition from a dirty context.