Add CommonScope
Adds another static scope named "CommonScope", accessible as
`config.commonScope` when defined in the application's scopes.js

The main reason to add this is the fact that is is sometimes nice to
have a scope that is *not* magical like all the other scopes except the
process scope are. In that they synchronize themselves to either the
database or localStorage or sessionStorage.

This is the new, non-magic, plain old observable scope. No sync, no
nothing.

Just like LocalScope and SessionScope the scope can have a potentially
promise-returning init() method.

In this case it gets called once after the common scope is created and
the promise returned is waited on before the application init method
from the *-startup.js is called.