WorkingSet: Add workingSet.load()
Adds workingSetLoad(type,id,source) as a convenience method to support
the most common working set workflow.

If we want to load a new object, we have to make sure that we don't
accidentally load another version of an entity we already have. And only
if the working set does not contain that entity, we want to load it.

workingSet.load() with a GraphQLQuery
-------------------------------------

    workingSet.load("Foo", id, Q_Foo).then( obj => ... )

The given GraphQLQuery must have the given type as payload type. Load
will construct the FilterDSL query and call the query and register
the result before returning it

workingSet.load() with a function
---------------------------------

As alternative, you can provide a load function that returns a Promise
with the referenced entity data

    workingSet.load("Foo", id, sourceFn).then( obj => ...)

sourceFn is called to provive the object wrapped in a promise, then load
registers the object and returns it again.