FKSelector: Support filtered Queries
Make sure to keep the conditions in the given query across all internal
queries including the <DataGrid/> in the modal.

To make things easy, we just use and modify the defaultVariables of the
given GraphQLQuery.

Two tests are failing for some reason. (Oddly, the "simple search" test
fails, but the more complex still works). Skipping those two tests for
now :\