FKSelector: Optimize search query
For the search query resulting from the user typing a name into the
embedded search field, we have modified the pageSize of the given config
and set it to zero so now, disabling pagination.

This could be a lot of data though, and all we are interested in is our
match and if it's unique, AKA if there are more than 1 results. We don't
really care if it's 2 or 3 or a 100.

So we might just as well set the pagination size to 2.