WorkingSet: Add onNextChange()
Adds a new method onNextChange( () => ... ) which registers a function
callback to called after the next working set change has been processed.

So if this method is called, e.g. in on onChange handler, one would
have the guruantee that the working set changes reflect the state
resulting from the change event.

We keep an internal queue in secret which we call and flush on every
_updateChanges.