Remove handleChange from useEffect
The uncommented line make zero sense in retrospect and could only possibly only be true for an FK Selector that displays its id to the user. That however is the very thing we don't want to do. The current theory is that the effect never or only very rarely occurred because form recreation was preventing the update.