Implement InteractiveQuery.separate
Adds a convenience separator for multi-query query documents.

This allows you to use efficient fetching of multiple queries as one
query in one request and then separate those queries into individual
queries so that they work with the interactive UI widgets etc.

Example transition:

    "test-partial": {
        to: CRUDList,
        action: t => {

            // We execute a multi-query document...
            return Q_FooMulti.execute({
                config: {},
                config2: {}
            }).then(
                result => {

                    // .. and then extract one iQuery document with a generated query containing only that iQuery
                    const foos = InteractiveQuery.separate(result, "foos")
                    scope.updateFoos(foos)
                }
            )
        }
    }

The static separate method takes the multi-query result and the name of
the variable to extract from that.

It returns an iQuery document with a _query attached that was generated
from the multi-query document definition.

This methods extends our _query format with a "selections" object that
contains all the field selections below "rows".