Refactor to decompileFilter
The existing decribeFilter function wasn't working right for components
and it wasn't really readable. So I wrote some indentation code and
while I was doing that it occured to me that I was like 90% done with a
expression decompiler.

This refactors describeFilter into decompileFilter() that returns
pretty-printed source code compatible with the FilterDSL and of course
also much better readable as debug output.