WorkingSet: Add experimental computed reducer
Our problem about wanting to do things with the current list of domain
objects resulting from a source list with all elements replaced with
those that are registered in the working set and all new objects of the
same type.

This adds a new experimentatl method

    workingSet.reducer(iQuery, (domainObject,v) => ... , <initial-value>)

iQuery is an iQuery document, the reducer function receives a
domain-object and the current reducer value. The last argument defines
the initial value for the reducer.

If this method is not invoked in a reactive context, it does not help
much, but it also shouldn't hurt much.

For a reactive context, e.g. you actually render the reducer result, it
should be more efficient, especially since we integrate the reducer
function into the computedFn, which means we will only update when the
fields the reducer accesses change, and not on every change.