Add support for headings in associationSelector
associationSelector can now optionally receive heading-strings from
inputSchema to support custom DataGrid Column headers.
------------------------------------------------------------------------
Add functionality to define visible columns

To be able to only show columns we want to be present in the
associationSelectorModal, the possibility to define visible columns
through two means has been implemented.

Normally all columns returned by the query are shown in the
associationSelectorModal, but if the input shema includes a type meta of
the name "associationSelectorVisibleColumns" it will be used to detemine
the columns that should be visible and discards the rest.

The property "visibleColumns" can be used to override the behaviour
introduced by the type meta "associationSelectorVisibleColumns".
Through this it will be possible, to test the new behaviour and, in case
it is needed, change the bahaviour of single associationSelector fields
without the need to introduce a new type.

Both definitions allow an array of the column names or a string with the
names as comma seperated values.

The binding strength is as follows:
No definition < type meta definition < property definition