Activate change marker for special cases
The change marker was not shown for AssociationSelector and DataGrid in
forms.
The old reference to section was stored inside the fieldContext which is
not used in special cases. To counter that the reference is now stored
in the dataset of the associated dom element.
For AssociationSelector this is the button element used to select
entries, for DataGrid the table itself is used.
To achieve this a new propery "name" was added to DataGrid which is
expected to reference the domainObject name.
This change to DataGrid ensures that adding or removing entries will
not break the change marker behaviour for shortcuts.

For AssociationSelector usage no changes have to be made.
DataGrids however need to add the new "name" property to fully enable
the feature:
<DataGrid ... name="{domainObject name}">...</DataGrid>