Add core drag and drop functionality to DataGrid
Add sortColumn attribute to DataGrid to enable sorting of rows in the
table.
The sort column will automatically apply the functionality to drag and
drop rows over each other to manually sort the data.

Curently only the visible part is implemented and the data is not
persisted, as the interface for that is not implemented yet.