Add support for custom filter alias
It is now possible to register custom filter for DataGrid as alias to
be interpreted instead of registering the same filter function multiple
times for many Columns. That way errors are less prone to happen and
the overall readability and usability of code is enhanced.

To use this the functionality registerCustomFilter in CustomFilter has
to be executed with the first parameter being the alias and the second
being filter function to be executed (or a built-in filter name).
The registerCustomFilter function is also exported by the index.
All filter fnctions will now have the fieldName associated with as first
parameter.