Add OfflineQuery and enable DataGrid array usage
Add OfflineQuery to enable paging, filtering and sorting preloaded data.
OfflineQueries can be generated with an InteractiveQuery or an Array as
base, where the latter will disable the filter functionality of DataGrid
as we can currently not determine the field definitions for the filters
without a base type.
DataGrid will automatically convert array data into OfflineQuery.

Allow usage of OfflineQuery in FKSelector.
OfflineQuery in FKSelector with array base data has to be tested.