Fix filter for null values and null error
Make it so that contains checks will not return true if the stack that
is compared is not set, instead retrun false.
Likewise for not contains checks but it was false and is now true.

Also do not trust values given and only lowercase if they are not null
or if it is a function only if the function returns a non null value.