Create static execute method in InteractiveQuery
To ensure that we do not loose track of conditions in InteractiveQuery
(such as filters) and still be able to run multiple queries in one
request, we need a method to pull all needed queries into one big query,
split the results and write them back into each respective query
instance.
This method is added and can be used to ensure we preserve all that
precious configurations while executing queries in a batch.

To use this simply call InteractiveQuery.updateAll and as parameters
either give a bunch of queries as parameters or objects containing
the query and an override config to manipulate the existing one.