Add triggerToastsForErrors util
Adds a common way to report the unified error messages we have now as
toasts.

The toasts have the advantage in many situations that they're not new
views that destroy and unmount the current view, but instead just notify
the user of problems and potentially recover from it.

We introduce one special magic error message "SESSION_EXPIRED" which the
server can use to signal that the session the user is in has expired for
whatever reason and thus the current state is lost and the user has to
reload (and potentially reauthenticate).

We use the normal autoClose feature for most errors but we make the user
close "SESSION_EXPIRED" messages manually, because that error both
shouldn't be ignored and doesn't go away.