Simplify internal error view
With the new error handling using toasts in many situations, we don't
have a lot of situations left where we use the internal ErrorView.

We simplify that error view to not use the process layout for what most
likely is a broken or not-initialized process state.

So we use a simple bootstrap container instead.