Improve error handling
A lot of times we would totally destroy the current app for what is
totally a recoverable error condition.

This commit changes the most error handling for the unified response
JSON users to display the errors via the new triggerToastsForErrors.

We make sure that we can recover from internet connectivity issues.