Add dependency resolver  and add it to DataGrid
DataGrid data might depend on other data based on views, etc.
To effectively resolve the data on local changes to be directly visible
inside the view without uploading, we add dependency metadata and a
resolver utility to check recursively if any dependencies can be found.
These dependencies are used to update any DataGrid with dependent data.