Add dynamic translated component
To ensure translations are properly applied, even if the component is
created before the translation is loaded, a dynamically translated
component is needed.
This new component will automatically update its translation if either
the value (translation key), the translation arguments, the renderer or
the translation itself changes.
The renderer is optionally added to change the resulting output, which
is simply the text by default.