Fix Expressions
The signature changes for renderLayoutNodes where not made for the
ExpressionDialog leading to errors.

This commits reverts the redundant addition of "rootType" to
renderLayoutNodes because it is already contained in the editor state

This method has way too many arguments. Maybe we should at least push
some of them to options? Introduce a QueryEditorContext?