Improvements for QueryEditor
To change the query editor to be more integrated, the onChange handler
is introduced.
This makes the save button optional. it will only be active, if the
saveButtonOnClick handler is set.

Additionally a data storage of any kind is now introduced to the query
editor, it can be an observable storage or a simple object.

The column selection is now causing a form error when it has no
selection.