Track request time
Implementing session timeout functionality seems to be far too
complicated to implement framework side. What we need to do to enable
is, is tracking when the framework does requests.

This adds a simple moduleutil/latestRequestInSession.js that stores the
time of the latest request as ISO-8601 string in localStorage. This
ensures that we can have multiple tabs / windows for the same session
and track expected session time out reliably.

registerRequestForSession() tracks the current time as latest request
time and latestRequestInSession() returns that time as DateTime.

Adds invocations of registerRequestForSession() for all framework
requests that do not end up in a network error.

Exports both methods from the library.