Implement ConditionPointer
Implements a specialized formal solution to an impromptu pattern.

ConditionPointer encapsulates an observable root object of a condition
node hierarchy and a lodash-y path to pointing to a node within that
hierarchy.

The <ConditionEditor/> uses this as an impromptu pattern, just like we
do for <Form> and <Field> components in domainql-form.

The ConditionPointer API now offers methods to deal with these pointers
and also can do assertions on operations to check that we're really
pointing to what we should point to etc.