Client-side impl of GraphQL exporting
Does some refactoring in InteractiveQuery to reuse things and adds a

    iQuery.export(exporter[, queryConfig])

method that by default calls the new GraphQLExportController to download
blob that then gets downloaded as dataURI.

The given exporter name must be matched with spring bean registration on
the server-side.

A queryConfig can be given as second paramter to control the export and
change the behavior from the default which is to use the current state
of the iQuery and turn off pagination for the export.

There also exists another entry point for exporting which is the static

    GraphQLQuery.export(exporter[, variables])

method. It allows for the export of multi-query documents. The second
paramter variables defaults to the static default variables of the
GraphQLQuery.

Adds a jsdoc typedef for our QueryConfig type.