Feature/link add description class (#149)
* Fix likePattern filter to only be applied if set

The filter value for the likePattern filter was generated even if its
field was empty. This is now fixed by returning a null value if the
filter is not set or empty. For the filter decompiler a filter value
template getter is added.

* Add a className to the attachments description

This is to help style the description to better fit into tight spaces.

---------

Co-authored-by: Denis Peters <denis.peters@quinscape.de>